refactor: replace OnFailureChain string with OnFailureChainSpec
Move OnFailureChainSpec from pkg/config to pkg/task (task-domain type).
Defaults.OnFailureChain and task.Spec.OnFailureChain now carry the
structured spec type. Bare-string YAML continues to parse correctly
via OnFailureChainSpec.UnmarshalYAML (Task 8).

This task only swaps the types — chain.Params merging into the
chained run's input is Task 10; reserved-key + autonomous-at-defaults
validation is Task 11.

Refs #236
