+++ pkg/task/onfailurechain.go
@@ -1,4 +1,4 @@
-package config
+package task
 
 import (
 	"fmt"
