+++ pkg/task/onfailurechain_test.go
@@ -1,4 +1,4 @@
-package config
+package task
 
 import (
 	"reflect"
